Reverted Translate system.
Reverted Translate system as I decided to give up on adding additional languages for the time being due to several issues I was having. I may tackle this again in the future
